% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowquery-package.R
\docType{package}
\name{snowquery-package}
\alias{snowquery}
\alias{snowquery-package}
\title{snowquery: Query 'Snowflake' Databases with 'SQL'}
\description{
A wrapper allowing 'SQL' queries to be run on a 'Snowflake' instance directly from an 'R' script, by using the 'snowflake-connector-python' package in the background.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mermelstein/snowquery}
  \item Report bugs at \url{https://github.com/mermelstein/snowquery/issues}
}

}
\author{
\strong{Maintainer}: Dani Mermelstein \email{dmermelstein@hey.com} [copyright holder]

}
\keyword{internal}
