% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDB.R
\name{queryDB}
\alias{queryDB}
\title{Query a Snowflake database}
\usage{
queryDB(
  query,
  db_type,
  username = NULL,
  password = NULL,
  host = NULL,
  port = NULL,
  database = NULL,
  warehouse = NULL,
  account = NULL,
  role = NULL
)
}
\arguments{
\item{query}{A string of the SQL query to execute}
}
\value{
A data frame containing the results of the query
}
\description{
Run a SQL query on a Snowflake database (requires a \verb{~/snowquery_creds.yaml} file)
}
\examples{
\dontrun{
# Query the database and get a dataframe of results
result <- queryDB("SELECT * FROM my_table")
print(result)
}

}
