% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowquery-package.R
\docType{package}
\name{snowquery-package}
\alias{snowquery}
\alias{snowquery-package}
\title{snowquery: Run SQL on Snowflake, Redshift, Postgres, SQLite and DuckDB}
\description{
Provides a unified interface to run SQL queries against Snowflake, Amazon Redshift, PostgreSQL, SQLite and DuckDB from R. Optionally streams large remote query results directly into a local DuckDB file for fast, reproducible, low-cost analytics. Snowflake queries are executed via the snowflake-connector-python through reticulate; the package auto-installs and enforces a minimum connector version. Credentials are managed via a single YAML file supporting multiple named connections.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mermelstein/snowquery}
  \item Report bugs at \url{https://github.com/mermelstein/snowquery/issues}
}

}
\author{
\strong{Maintainer}: Dani Mermelstein \email{dmermelstein@hey.com} [copyright holder]

}
\keyword{internal}
