% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowquery-package.R
\docType{package}
\name{snowquery-package}
\alias{snowquery}
\alias{snowquery-package}
\title{snowquery: SQL Interface to 'Snowflake', 'Redshift', 'Postgres', 'SQLite', and 'DuckDB'}
\description{
Run 'SQL' queries across 'Snowflake', 'Amazon Redshift', 'PostgreSQL', 'SQLite', and 'DuckDB' from R with a single function. Optionally stream and cache large query results to a local 'DuckDB' database for efficient work with larger-than-memory datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mermelstein/snowquery}
  \item Report bugs at \url{https://github.com/mermelstein/snowquery/issues}
}

}
\author{
\strong{Maintainer}: Dani Mermelstein \email{dmermelstein@hey.com} [copyright holder]

}
\keyword{internal}
